+++ sdk/identity/identity/src/credentials/clientSecretCredential.ts
@@ -39,7 +39,7 @@ export class ClientSecretCredential implements TokenCredential {
     options: ClientSecretCredentialOptions = {}
   ) {
     if (!tenantId || !clientId || !clientSecret) {
-      throw new Error("tenantId, clientId, and clientSecret are required parameters.");
+      throw new Error("ClientSecretCredential: tenantId, clientId, and clientSecret are required parameters.");
     }
     this.msalFlow = new MsalClientSecret({
       ...options,